good night

+++ luckynum.c
@@ -9,7 +9,7 @@ int main()
     if (num >= 10 && num <= 99)
     {
         reNum = num % 10;
-        quNum = (int)num / 10;
+        quNum = num / 10;
         // printf("reNum = %d, quNum = %d", reNum, quNum);
 
         if (reNum % quNum == 0 || quNum % reNum == 0)
